refactor(web): rename tools interfaces.ts to types.ts
Matches the majority convention in lib/ (13 types.ts against 8 interfaces.ts)
and the sibling feature directories this code is read alongside - agents,
projects, settings.

Also switches the file's three relative imports to the @ prefix the repo
standard requires; they were only reachable as relative because the importers
sit in the same directory.
