+++ web/src/app/craft/v1/apps/connectableApps.ts
@@ -17,7 +17,7 @@ import {
   MCPAuthenticationPerformer,
   MCPAuthenticationType,
   MCPServer,
-} from "@/lib/tools/interfaces";
+} from "@/lib/tools/types";
 import { getActionIcon } from "@/lib/tools/mcpUtils";
 import { CRAFT_APPS_PATH } from "@/app/craft/v1/constants";
