+++ web/src/hooks/useAvailableTools.ts
@@ -1,7 +1,7 @@
 "use client";
 
 import useSWR from "swr";
-import { ToolSnapshot } from "@/lib/tools/interfaces";
+import { ToolSnapshot } from "@/lib/tools/types";
 import { errorHandlingFetcher } from "@/lib/fetcher";
 import { SWR_KEYS } from "@/lib/swr-keys";
