+++ web/src/hooks/useServerTools.ts
@@ -3,7 +3,7 @@
 import useSWR, { KeyedMutator } from "swr";
 import { errorHandlingFetcher } from "@/lib/fetcher";
 import { getActionIcon } from "@/lib/tools/mcpUtils";
-import { MCPServer, MCPTool, ToolSnapshot } from "@/lib/tools/interfaces";
+import { MCPServer, MCPTool, ToolSnapshot } from "@/lib/tools/types";
 
 /**
  * Return type for the useServerTools hook
