+++ web/src/lib/tools/mcpUtils.tsx
@@ -1,5 +1,5 @@
-import { SOURCE_METADATA_MAP } from "../sources";
-import { MCPServer } from "./interfaces";
+import { SOURCE_METADATA_MAP } from "@/lib/sources";
+import { MCPServer } from "@/lib/tools/types";
 import type { IconProps } from "@opal/types";
 import { SvgFileText, SvgServer } from "@opal/icons";
