+++ web/src/sections/actions/ActionCardHeader.tsx
@@ -2,7 +2,7 @@
 
 import React, { useState } from "react";
 import { cn } from "@opal/utils";
-import { ActionStatus } from "@/lib/tools/interfaces";
+import { ActionStatus } from "@/lib/tools/types";
 import Text from "@/refresh-components/texts/Text";
 import IconButton from "@/refresh-components/buttons/IconButton";
 import ButtonRenaming from "@/refresh-components/buttons/ButtonRenaming";
